Добавление возможности покупки ценных бумаг и связывание с БД. Реализация необходимых средств взаимодействия с БД для покупки ценных бумаг и рефакторинг этого взаимодействия

+++ SimpleTrader.Domain/Models/Asset.cs
@@ -3,6 +3,6 @@
     public class Asset
     {
         public string Symbol { get; set; }
-        public decimal PricePerShare { get; set; }
+        public double PricePerShare { get; set; }
     }
 }
